+++ core/utils/utils.go
@@ -104,82 +104,6 @@ func IsEmpty(bytes []byte) bool {
 	return true
 }
 
-// Sleeper interface is used for tasks that need to be done on some
-// interval, excluding Cron, like reconnecting.
-type Sleeper interface {
-	Reset()
-	Sleep()
-	After() time.Duration
-	Duration() time.Duration
-}
-
-// BackoffSleeper is a sleeper that backs off on subsequent attempts.
-type BackoffSleeper struct {
-	backoff.Backoff
-	beenRun atomic.Bool
-}
-
-// NewBackoffSleeper returns a BackoffSleeper that is configured to
-// sleep for 0 seconds initially, then backs off from 1 second minimum
-// to 10 seconds maximum.
-func NewBackoffSleeper() *BackoffSleeper {
-	return &BackoffSleeper{
-		Backoff: backoff.Backoff{
-			Min: 1 * time.Second,
-			Max: 10 * time.Second,
-		},
-	}
-}
-
-// Sleep waits for the given duration, incrementing the back off.
-func (bs *BackoffSleeper) Sleep() {
-	if bs.beenRun.CompareAndSwap(false, true) {
-		return
-	}
-	time.Sleep(bs.Backoff.Duration())
-}
-
-// After returns the duration for the next stop, and increments the backoff.
-func (bs *BackoffSleeper) After() time.Duration {
-	if bs.beenRun.CompareAndSwap(false, true) {
-		return 0
-	}
-	return bs.Backoff.Duration()
-}
-
-// Duration returns the current duration value.
-func (bs *BackoffSleeper) Duration() time.Duration {
-	if !bs.beenRun.Load() {
-		return 0
-	}
-	return bs.ForAttempt(bs.Attempt())
-}
-
-// Reset resets the backoff intervals.
-func (bs *BackoffSleeper) Reset() {
-	bs.beenRun.Store(false)
-	bs.Backoff.Reset()
-}
-
-// RetryWithBackoff retries the sleeper and backs off if not Done
-func RetryWithBackoff(ctx context.Context, fn func() (retry bool)) {
-	sleeper := NewBackoffSleeper()
-	sleeper.Reset()
-	for {
-		retry := fn()
-		if !retry {
-			return
-		}
-
-		select {
-		case <-ctx.Done():
-			return
-		case <-time.After(sleeper.After()):
-			continue
-		}
-	}
-}
-
 // UnmarshalToMap takes an input json string and returns a map[string]interface i.e. a raw object
 func UnmarshalToMap(input string) (map[string]interface{}, error) {
 	var output map[string]interface{}
